Different norm_cer calculation, bug fix
We calculate the normalized CER using the normalized GT and the
NORMAL prediction. We do this so we can see the difference in
performance between using the norm file and not using it. We are
not interested in altering the output of the model, since that is
to be determined BEFORE training